Rename explorer directory to home
We no longer have any Explorer component so this naming is obsolete.
