+++ app/src/app/components/contents/home/PacketGroupSummaryListItem.tsx
