+++ app/src/app/components/contents/home/hooks/useGetPacketGroupSummary.ts
