+++ app/src/tests/components/contents/home/Home.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { MemoryRouter } from "react-router";
 import { SWRConfig } from "swr";
-import { Home } from "../../../../app/components/contents/explorer";
+import { Home } from "../../../../app/components/contents/home";
 import * as fetch from "../../../../lib/fetch";
 
 describe("Home component", () => {
