+++ app/src/tests/components/login/Redirect.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import { MemoryRouter, Route, Routes } from "react-router-dom";
-import { Home } from "../../../app/components/contents/explorer";
+import { Home } from "../../../app/components/contents/home";
 import { Login, Redirect } from "../../../app/components/login";
 import { UserState } from "../../../app/components/providers/types/UserTypes";
 import { mockUserState, mockExpiredUserState } from "../../mocks";
